Fix warnings in sample_consensus/sac_model_circle3d.hpp
> sac_model_circle3d.hpp:283:24: warning: comparison of integers of different signs: 'size_t' (aka 'unsigned long') and 'Index' (aka 'long') [-Wsign-compare]

+++ sample_consensus/include/pcl/sample_consensus/impl/sac_model_circle3d.hpp
@@ -280,7 +280,7 @@ pcl::SampleConsensusModelCircle3D<PointT>::optimizeModelCoefficients (
   Eigen::LevenbergMarquardt<Eigen::NumericalDiff<OptimizationFunctor>, double> lm (num_diff);
   Eigen::VectorXd coeff;
   int info = lm.minimize (coeff);
-  for (size_t i = 0; i < coeff.size (); ++i)
+  for (int i = 0; i < coeff.size (); ++i)
     optimized_coefficients[i] = static_cast<float> (coeff[i]);
 
   // Compute the L2 norm of the residuals
